feat(ui): new option to grey completed/entered habits
Works the same way as hiding habits, but only changes the color.

The reason this is useful, is because when a habit is treated as
completed/entered, the cause might be just the frequency set for
that habit. So, if it's once per month and you've entered it only
once, till a month has passed the habit will be hidden, but
the habit might've been done again within that month and it would
be useful to be able to see it to enter that data.

At some point it'll be good to change the darkest grey in the
habit palet in "Edit habit", so that it doesn't conflict with
the archived habit and "grey complete" colors.