fix(ui): hacky fix for updating colors on show question marks setting change
This only applies if the grey completed setting is enabled.
This is to avoid restarting the app to see the resulting color change,
when changing the show question marks setting.

It'll be good to call `finish()` on the settings activity,
to avoid stacking activities every time the setting is changed.